+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this project will be documented in this file.
 
+# Unreleased
+- add `DuckDB::Connection#prepare`. `DuckDB::Connection#prepare` is an alias of `DuckDB::Connection#prepared_statement`.
+
 # 1.1.0.0 - 2024-09-15
 - drop ruby 3.0.x.
 - bump duckdb to 1.1.0.